pyproject.toml
[project]
name = "idea-explorer"
version = "0.1.0"
description = "Autonomous research framework that takes structured research ideas and orchestrates AI agents to design, execute, analyze, and document experiments across diverse domains"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "ChicagoHAI" }
]
keywords = ["research", "ai-agents", "automation", "scientific-computing", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "PyYAML>=6.0.1",
    "Jinja2>=3.1.2",
    "PyGithub>=2.4.0",
    "GitPython>=3.1.43",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/ChicagoHAI/idea-explorer"
Documentation = "https://github.com/ChicagoHAI/idea-explorer/blob/main/README.md"
Repository = "https://github.com/ChicagoHAI/idea-explorer"
Issues = "https://github.com/ChicagoHAI/idea-explorer/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
